backend/requirements.txt
pandas
numpy
scikit-learn
tensorflow
prophet
yfinance
fastapi
uvicorn
joblib
plotly
mysql-connector-python